+++ tests/collection/collection-layout.spec.ts
@@ -60,7 +60,7 @@ test(`Sort by All-time views in Tile view`, async ({ collectionPage }) => {
       'descending',
       10,
     );
-    await collectionPage.searchPage.validateURLParamsWithSortFilter(
+    await collectionPage.collectionBrowser.validateURLParamsWithSortFilter(
       'All-time views',
       'descending',
     );
@@ -85,7 +85,7 @@ test(`Sort by Date published in List view`, async ({ collectionPage }) => {
       'descending',
       10,
     );
-    await collectionPage.searchPage.validateURLParamsWithSortFilter(
+    await collectionPage.collectionBrowser.validateURLParamsWithSortFilter(
       'Date published',
       'descending',
     );
@@ -107,10 +107,10 @@ test(`Sort by Date archived (ascending) in Compact view`, async ({
   });
 
   await test.step('Check list column headers for sort filter', async () => {
-    await collectionPage.searchPage.validateCompactViewModeListLineDateHeaders(
+    await collectionPage.collectionBrowser.validateCompactViewModeListLineDateHeaders(
       'Date archived',
     );
-    await collectionPage.searchPage.validateURLParamsWithSortFilter(
+    await collectionPage.collectionBrowser.validateURLParamsWithSortFilter(
       'Date archived',
       'ascending',
     );
@@ -124,9 +124,9 @@ test.beforeEach(async ({ collectionPage }) => {
   });
 
   await test.step(`Select "Search metadata" and do a metadata search for "radio"`, async () => {
-    await collectionPage.searchPage.clickSearchInputOption(
+    await collectionPage.collectionBrowser.clickSearchInputOption(
       SearchOption.METADATA,
     );
-    await collectionPage.searchPage.queryFor('radio');
+    await collectionPage.collectionBrowser.queryFor('radio');
   });
 });
