+++ tests/collection/collection-search.spec.ts
@@ -3,13 +3,13 @@ import { SearchOption } from '../models';
 
 test('Collection search metadata', async ({ collectionPage }) => {
   await test.step(`Select "Search metadata"`, async () => {
-    await collectionPage.searchPage.clickSearchInputOption(
+    await collectionPage.collectionBrowser.clickSearchInputOption(
       SearchOption.METADATA,
     );
   });
 
   await test.step(`Search for "radio" in the search input text field`, async () => {
-    await collectionPage.searchPage.queryFor('radio');
+    await collectionPage.collectionBrowser.queryFor('radio');
   });
 
   await test.step(`Results are displayed in display area - validate first item displayed`, async () => {
@@ -22,11 +22,13 @@ test('Collection search text contents and clear filters', async ({
   collectionPage,
 }) => {
   await test.step(`Select "Search text contents"`, async () => {
-    await collectionPage.searchPage.clickSearchInputOption(SearchOption.TEXT);
+    await collectionPage.collectionBrowser.clickSearchInputOption(
+      SearchOption.TEXT,
+    );
   });
 
   await test.step(`Search for "dragnet" in the search input text field`, async () => {
-    await collectionPage.searchPage.queryFor('dragnet');
+    await collectionPage.collectionBrowser.queryFor('dragnet');
   });
 
   await test.step(`Results are displayed in display area - validate first item displayed`, async () => {
@@ -35,22 +37,22 @@ test('Collection search text contents and clear filters', async ({
   });
 
   await test.step(`Click "X" button in search input and validate search input text is cleared`, async () => {
-    await collectionPage.searchPage.clickClearSearchInput();
-    await collectionPage.searchPage.validateClearSearchInput();
+    await collectionPage.collectionBrowser.clickClearSearchInput();
+    await collectionPage.collectionBrowser.validateClearSearchInput();
   });
 });
 
 test('No results page displays when no results', async ({ collectionPage }) => {
   await test.step(`Select "Search metadata"`, async () => {
-    await collectionPage.searchPage.clickSearchInputOption(
+    await collectionPage.collectionBrowser.clickSearchInputOption(
       SearchOption.METADATA,
     );
   });
 
   await test.step(`Search for "catsshfksahfkjhfkjsdhfkiewhkdsfahkjhfkjsda" and validate that the "No results" placeholder appears in place of the display area`, async () => {
-    await collectionPage.searchPage.queryFor(
+    await collectionPage.collectionBrowser.queryFor(
       'catsshfksahfkjhfkjsdhfkiewhkdsfahkjhfkjsda',
     );
-    await collectionPage.searchPage.validateEmptyPagePlaceholder();
+    await collectionPage.collectionBrowser.validateEmptyPagePlaceholder();
   });
 });
